back/requirements.txt
Django==3.2.2
asgiref==3.3.4
django-filter==2.4.0
django-restql==0.11.2
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev2
pip==21.1.1
psycopg2==2.8.6
pyPEG2==2.15.2
pytz==2021.1
setuptools==56.1.0
six==1.16.0
sqlparse==0.4.1
django-cors-headers==3.7.0